new nuget examples
