+++ .gitignore
@@ -4,5 +4,7 @@
 bin/
 obj/
 .DS_Store
-assets/output.png
 *.snupkg
+
+output_image.png
+output_file.pdf
