--- examples/image-to-pdf/image-to-pdf.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26124.0
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ImageToPdf", "ImageToPdf\ImageToPdf.csproj", "{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Debug|x64.Build.0 = Debug|Any CPU
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Debug|x86.Build.0 = Debug|Any CPU
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Release|x64.ActiveCfg = Release|Any CPU
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Release|x64.Build.0 = Release|Any CPU
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Release|x86.ActiveCfg = Release|Any CPU
-		{C3E4D765-A743-425B-8BF5-08D3F9DC72A8}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FAD06ADA-2CBB-4AC3-977C-8FB53112F142}
-	EndGlobalSection
-EndGlobal
