+++ examples/nuget-usage/NugetUsageX64/NugetUsageX64.csproj
@@ -7,6 +7,7 @@
 
   <ItemGroup>
     <PackageReference Include="Docnet.Core" Version="2.1.0" />
+    <PackageReference Include="System.Drawing.Common" Version="4.5.1" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.1" />
     <PackageReference Include="xunit" Version="2.4.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.2">
@@ -16,7 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Content Include="..\Assets\image_0.png" Link="Assets\image_0.png">
+    <Content Include="..\Assets\image_0.jpeg" Link="Assets\image_0.jpeg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="..\Assets\wikipedia_0.pdf" Link="Assets\wikipedia_0.pdf">
