--- examples/nuget-usage/NugetUsageX64/Program.cs
@@ -1,5 +1,0 @@
-﻿using System;
-
-namespace NugetUsageX64
-{
-}
