+++ examples/nuget-usage/NugetUsageX86/NugetUsageX86.csproj
@@ -7,8 +7,13 @@
 
   <ItemGroup>
     <PackageReference Include="Docnet.Core" Version="2.1.0" />
+    <PackageReference Include="System.Drawing.Common" Version="4.5.1" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.1" />
     <PackageReference Include="xunit" Version="2.4.1" />
+    <PackageReference Include="xunit.runner.console" Version="2.4.1">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.2">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
@@ -16,7 +21,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Content Include="..\Assets\image_0.png" Link="Assets\image_0.png">
+    <Content Include="..\Assets\image_0.jpeg" Link="Assets\image_0.jpeg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="..\Assets\wikipedia_0.pdf" Link="Assets\wikipedia_0.pdf">
