--- examples/nuget-usage/NugetUsageX86/Program.cs
@@ -1,5 +1,0 @@
-﻿using System;
-
-namespace NugetUsageX86
-{
-}
